--- a/mlserver/parallel.py
+++ b/mlserver/parallel.py
@@ -1,3 +1,4 @@
+import time
 import asyncio
 import multiprocessing as mp
 
@@ -10,6 +11,7 @@ from .settings import ModelSettings
 from .model import MLModel
 from .types import InferenceRequest, InferenceResponse
 from .utils import get_wrapped_method
+from .logging import logger
 
 _InferencePoolAttr = "__inference_pool__"
 
@@ -87,6 +89,21 @@ class InferencePool:
         loop = asyncio.get_running_loop()
         return await loop.run_in_executor(self._executor, _mp_predict, payload)
 
+    async def warm_workers(self, model: MLModel):
+        logger.info(
+            f"Warm workers enabled - loading {model.name} in inference pool workers"
+        )
+        loop = asyncio.get_event_loop()
+
+        # Submit tasks to trigger initializer
+        loading_tasks = [
+            loop.run_in_executor(self._executor, _noop)
+            for _ in range(model.settings.parallel_workers)
+        ]
+        logger.info("Waiting for models to be loaded")
+
+        return await asyncio.wait(loading_tasks)
+
     def __del__(self):
         self._executor.shutdown(wait=True)
 
@@ -130,6 +147,12 @@ async def load_inference_pool(model: MLModel):
     # Decorate predict method
     setattr(model, "predict", parallel(model.predict))
 
+    # Conditionally load models to all workers in thread pool executor
+    # This will cut down on initial inference response times but
+    # will increase the amount of RAM utilized
+    if model.settings.warm_workers == True:
+        await pool.warm_workers(model)
+
     return model
 
 
@@ -140,3 +163,8 @@ async def unload_inference_pool(model: MLModel):
 
     pool.__del__()
     delattr(model, _InferencePoolAttr)
+
+
+def _noop():
+    time.sleep(0.1)
+    return None